Add files via upload
There's nothing interactive. I used only JavaScript because, well, I'm an amateur for the time being. This app is still something I built myself and it's the first app I ever built so it's not the best. Still, anyone's welcome to look at it. Hopefully someone can see the underlying code and if not that's fine too. In any case, I only just finished a Codecademy course on JavaScript and I'd say it's pretty good considering I have nobody to teach me yet.